benchmarks: dissolve PR regression shards into per-task leaves
Rename and split remaining test_task_regressions_* dumps into task-owned
modules, fold overlapping core/287 cases into existing leaves, drop
redundant reference smokes already covered by RESOURCE_DERIVED oracle
checks or alternate-accept suites, and document RESOURCE_DERIVED as a
curated sample.
