+++ benchmarks/validation/agent_workflow_v1/test_autoformalization_semantic_audit.py
@@ -3,9 +3,44 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
+import pytest
 from benchmarks.validation.agent_workflow_v1 import support
 
 TASK = "autoformalization-semantic-audit"
+
+
+def test_accepts_alternative_exact_witnesses(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["missing_premise_certificate"] = {
+        "dimension": 1,
+        "x": [-7],
+        "forced_y": [0],
+    }
+    submission["result"]["operator_mismatch_certificate"] = {
+        "dimension": 2,
+        "x": [3, -2],
+        "y": [2, 3],
+        "dot_product": 0,
+        "coordinate_products": [6, -6],
+    }
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+    accepted = support._run_verifier(task, app, logs)
+    assert accepted["correctness"] == 1.0
+    assert accepted["reward"] == pytest.approx(1.0)
+
+
+def test_rejects_incomplete_defect_set(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["defects"] = ["MISSING_DIMENSION_PREMISE"]
+    support._write_json(app / "submission.json", submission)
+
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
+    assert rejected["reward"] == 0.0
 
 
 def test_rejects_positive_lean_compile_claim(tmp_path: Path) -> None:
