+++ benchmarks/validation/agent_workflow_v1/test_finite_magma_countermodel.py
@@ -3,9 +3,22 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
+import pytest
 from benchmarks.validation.agent_workflow_v1 import support
 
 TASK = "finite-magma-countermodel"
+
+
+def test_accepts_alternate_smallest_countermodel(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["table"] = [[1, 0], [1, 0]]
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+    accepted = support._run_verifier(task, app, logs)
+    assert accepted["correctness"] == 1.0
+    assert accepted["reward"] == pytest.approx(1.0)
 
 
 def test_rejects_corrupted_table(tmp_path: Path) -> None:
