+++ benchmarks/validation/agent_workflow_v1/test_noncompact_lefschetz_proof_audit.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 from benchmarks.validation.agent_workflow_v1 import support
 
-TASK = "metric-tsp-proof-repair"
+TASK = "noncompact-lefschetz-proof-audit"
 
 
 def _case(tmp_path: Path):
@@ -18,32 +18,34 @@ def _rewrite(app: Path, submission: dict) -> None:
     support._write_json(app / "submission.json", submission)
 
 
-def test_accepts_factor_two_claim(tmp_path: Path) -> None:
+def test_accepts_equivalent_rational_and_cohomology_forms(tmp_path: Path) -> None:
     task, app, logs = _case(tmp_path)
     submission = json.loads((app / "submission.json").read_text())
-    submission["result"]["corrected_claim"] = "factor-2 approximation"
+    counterexample = submission["result"]["counterexample"]
+    counterexample["translation"] = {"numerator": 2, "denominator": 2}
+    counterexample["compact_support_cohomology"].reverse()
     _rewrite(app, submission)
-
     accepted = support._run_verifier(task, app, logs)
-    assert accepted["correctness"] == 1.0
-
-
-def test_accepts_reversed_optimal_tour(tmp_path: Path) -> None:
-    task, app, logs = _case(tmp_path)
-    submission = json.loads((app / "submission.json").read_text())
-    submission["result"]["optimal_tour"] = ["A", "D", "C", "F", "E", "B", "A"]
-    _rewrite(app, submission)
-
-    accepted = support._run_verifier(task, app, logs)
-    assert accepted["correctness"] == 1.0
     assert accepted["reward"] == pytest.approx(1.0)
 
 
-def test_rejects_corrupted_optimal_weight(tmp_path: Path) -> None:
+@pytest.mark.parametrize("field", ["top_degree_action", "lefschetz_number"])
+def test_rejects_boolean_in_integer_fields(tmp_path: Path, field: str) -> None:
     task, app, logs = _case(tmp_path)
     submission = json.loads((app / "submission.json").read_text())
-    submission["result"]["weights"]["optimal"] = 31
-    support._write_json(app / "submission.json", submission)
+    submission["result"]["counterexample"][field] = True
+    _rewrite(app, submission)
     rejected = support._run_verifier(task, app, logs)
     assert rejected["correctness"] == 0.0
-    assert rejected["reward"] == 0.0
+
+
+def test_enforces_visible_translation_bounds(tmp_path: Path) -> None:
+    task, app, logs = _case(tmp_path)
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["counterexample"]["translation"] = {
+        "numerator": 1_000_001,
+        "denominator": 1,
+    }
+    _rewrite(app, submission)
+    rejected = support._run_verifier(task, app, logs)
+    assert rejected["correctness"] == 0.0
