+++ benchmarks/validation/agent_workflow_v1/test_symmetric_polynomial_divisibility.py
@@ -5,31 +5,23 @@ from pathlib import Path
 
 from benchmarks.validation.agent_workflow_v1 import support
 
-TASK = "polynomial-tail-counterexample"
+TASK = "symmetric-polynomial-divisibility"
 
 
-def test_rejects_non_array_witness_fields(tmp_path: Path) -> None:
+def test_rejects_corrupted_multiplier(tmp_path: Path) -> None:
     task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
     submission = json.loads((app / "submission.json").read_text())
-    submission["result"] = {
-        "p_coefficients": {"2": None, "8/3": None, "2/3": None, "0": None},
-        "q_coefficients": {"1": None, "199": None, "9900": None},
-        "p_roots": {"-1": None, "-1/3": None, "0": None},
-        "q_roots": {"-100": None, "-99": None},
-        "x1": "0",
-        "x2": "1/100",
-    }
+    submission["result"]["generator_multipliers"][0][0]["coefficient"] = "2"
     support._write_json(app / "submission.json", submission)
-
     rejected = support._run_verifier(task, app, logs)
     assert rejected["correctness"] == 0.0
     assert rejected["reward"] == 0.0
 
 
-def test_rejects_corrupted_x2(tmp_path: Path) -> None:
+def test_rejects_incomplete_multiplier(tmp_path: Path) -> None:
     task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
     submission = json.loads((app / "submission.json").read_text())
-    submission["result"]["x2"] = "1"
+    submission["result"]["generator_multipliers"][1].pop()
     support._write_json(app / "submission.json", submission)
     rejected = support._run_verifier(task, app, logs)
     assert rejected["correctness"] == 0.0
