--- benchmarks/validation/agent_workflow_v1/test_task_regressions_335_338.py
@@ -1,32 +1,0 @@
-from __future__ import annotations
-
-import json
-from pathlib import Path
-
-from benchmarks.validation.agent_workflow_v1 import support
-
-
-def test_exponential_moment_rejects_corrupted_generic_formula(tmp_path: Path) -> None:
-    task, app, logs = support._prepare_case(
-        tmp_path, "exponential-moment-rationality", "computed"
-    )
-    path = app / "submission.json"
-    submission = json.loads(path.read_text())
-    submission["result"]["generic_formula"]["numerator"][0]["coefficient"] = "2"
-    support._write_json(path, submission)
-    rejected = support._run_verifier(task, app, logs)
-    assert rejected["correctness"] == 0.0
-    assert rejected["reward"] == 0.0
-
-
-def test_exponential_moment_rejects_missing_singular_branch(tmp_path: Path) -> None:
-    task, app, logs = support._prepare_case(
-        tmp_path, "exponential-moment-rationality", "computed"
-    )
-    path = app / "submission.json"
-    submission = json.loads(path.read_text())
-    submission["result"]["singular_formula"] = submission["result"]["generic_formula"]
-    support._write_json(path, submission)
-    rejected = support._run_verifier(task, app, logs)
-    assert rejected["correctness"] == 0.0
-    assert rejected["reward"] == 0.0
