+++ benchmarks/validation/agent_workflow_v1/test_well_total_domination_counterexample.py
@@ -3,9 +3,26 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
+import pytest
 from benchmarks.validation.agent_workflow_v1 import support
 
 TASK = "well-total-domination-counterexample"
+
+
+def test_accepts_reordered_exact_witnesses(tmp_path: Path) -> None:
+    task, app, logs = support._prepare_case(tmp_path, TASK, "computed")
+    submission = json.loads((app / "submission.json").read_text())
+    submission["result"]["pendant_vertices"] = ["4", "0"]
+    submission["result"]["minimal_total_dominating_sets"] = [
+        ["4", "3", "1", "0"],
+        ["3", "2", "1"],
+    ]
+    support._bind_result_evidence(app, submission)
+    support._write_json(app / "submission.json", submission)
+
+    accepted = support._run_verifier(task, app, logs)
+    assert accepted["correctness"] == 1.0
+    assert accepted["reward"] == pytest.approx(1.0)
 
 
 def test_rejects_corrupted_degree_sum(tmp_path: Path) -> None:
